pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-python-boilerplate"
dynamic = ["version"]
description = "A comprehensive LLM inference toolkit with multi-provider support and tool integration"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Alessio", email = "your.email@example.com"}
]
maintainers = [
    {name = "Alessio", email = "your.email@example.com"}
]
keywords = ["llm", "inference", "openai", "aws", "bedrock", "azure", "tools", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "boto3>=1.26.0",
    "httpx>=0.24.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
full = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "pyttsx3>=2.90",
]

[project.urls]
Homepage = "https://github.com/AlessioGrespi/LLM-Python-Boilerplate"
Documentation = "https://github.com/AlessioGrespi/LLM-Python-Boilerplate#readme"
Repository = "https://github.com/AlessioGrespi/LLM-Python-Boilerplate"
"Bug Tracker" = "https://github.com/AlessioGrespi/LLM-Python-Boilerplate/issues"



[tool.setuptools.dynamic]
version = {attr = "ultimate_llm_toolkit.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
